+++ SampleCode/PingOneVerify/app/src/main/java/com/pingidentity/sdk/pingoneverify/ui/fragments/LocationFragment.kt
@@ -28,9 +28,9 @@ class LocationFragment(
     private lateinit var binding: DialogLocationBinding
 
 
-    private val mTimerObserver = object : DocumentSubmissionTimer.TimerObserver {
+    private val mTimerObserver = object : DocumentSubmissionTimer.Observer {
         override fun onTick(millisRemaining: Int) {
-            binding?.txtTimeRemaining?.text = getTimeRemainingString(millisRemaining, R.string.sdk_timer_label)
+            binding?.txtTimeRemaining?.text = getTimeRemainingString(millisRemaining, R.string.idv_timer_label)
         }
         override fun onFinish() {}
     }